refactor(date): update date imports to use dateUtils module
